Removing legacy AddTaskUseCase implementation.
