+++ app/src/main/java/com/adammcneilly/toa/addtask/ui/AddTaskViewModel.kt
@@ -6,7 +6,6 @@ import androidx.lifecycle.viewModelScope
 import com.adammcneilly.toa.R
 import com.adammcneilly.toa.addtask.domain.model.AddTaskResult
 import com.adammcneilly.toa.addtask.domain.model.TaskInput
-import com.adammcneilly.toa.addtask.domain.usecases.AddTaskUseCase
 import com.adammcneilly.toa.core.models.Task
 import com.adammcneilly.toa.core.ui.UIText
 import com.adammcneilly.toa.destinations.AddTaskScreenDestination
