+++ app/src/main/java/com/adammcneilly/toa/core/di/UseCaseModule.kt
@@ -1,7 +1,5 @@
 package com.adammcneilly.toa.core.di
 
-import com.adammcneilly.toa.addtask.domain.usecases.AddTaskUseCase
-import com.adammcneilly.toa.addtask.domain.usecases.ProdAddTaskUseCase
 import com.adammcneilly.toa.tasklist.domain.usecases.GetAllTasksUseCase
 import com.adammcneilly.toa.tasklist.domain.usecases.GetTasksForDateUseCase
 import com.adammcneilly.toa.tasklist.domain.usecases.ProdGetAllTasksUseCase
