+++ app/src/test/java/com/adammcneilly/toa/addtask/ui/AddTaskViewModelRobot.kt
@@ -3,7 +3,9 @@ package com.adammcneilly.toa.addtask.ui
 import androidx.lifecycle.SavedStateHandle
 import com.adammcneilly.toa.addtask.domain.model.AddTaskResult
 import com.adammcneilly.toa.core.models.Task
-import com.adammcneilly.toa.fakes.FakeAddTaskUseCase
+import com.adammcneilly.toa.fakes.FakePreferences
+import com.adammcneilly.toa.preferences.UserPreferences
+import com.adammcneilly.toa.task.api.test.FakeTaskRepository
 import com.google.common.truth.Truth.assertThat
 import io.mockk.every
 import io.mockk.mockk
@@ -12,13 +14,15 @@ import java.time.LocalDate
 import java.time.ZoneId
 
 class AddTaskViewModelRobot {
-    private val fakeAddTaskUseCase = FakeAddTaskUseCase()
+    private val fakeTaskRepository = FakeTaskRepository()
+    private val fakeUserPreferences = UserPreferences(FakePreferences())
     private val mockSavedStateHandle: SavedStateHandle = mockk(relaxed = true)
     private lateinit var viewModel: AddTaskViewModel
 
     fun buildViewModel() = apply {
         viewModel = AddTaskViewModel(
-            addTaskUseCase = fakeAddTaskUseCase.mock,
+            taskRepository = fakeTaskRepository,
+            userPreferences = fakeUserPreferences,
             savedStateHandle = mockSavedStateHandle,
         )
     }
@@ -31,11 +35,17 @@ class AddTaskViewModelRobot {
         } returns date
     }
 
+    /**
+     * NOTE: Since we're no longer depending on use case as its own type,
+     * we need to split this up into two functions: one mocks a response from the taskrepository
+     * and one mocks the response from the user preferences
+     */
     fun mockResultForTask(
         task: Task,
         result: AddTaskResult
     ) = apply {
-        fakeAddTaskUseCase.mockResultForTask(task, result)
+        TODO()
+//        fakeTaskRepository.addTaskResults[task] = result
     }
 
     fun enterDescription(
